Improve extensibility for commands
