+++ applications/lokqlDx/CommandProcessorProvider.cs
@@ -1,0 +1,16 @@
+﻿using Lokql.Engine.Commands;
+
+namespace lokqlDx;
+
+/// <summary>
+///     Indirection class to allow cleaner source-level extension of the CommandProcessor
+/// </summary>
+public static class CommandProcessorProvider
+{
+    public static CommandProcessor GetCommandProcessor()
+    {
+        return CommandProcessor.Default()
+            //Add .WithAdditionalCommand lines here to extend the processor
+            ;
+    }
+}
